# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DJERDJERA Comptable API - Algerian small-business accounting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /company:
    get:
      operationId: getCompany
      tags: [company]
      summary: Get the active company profile (single tenant)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
    put:
      operationId: upsertCompany
      tags: [company]
      summary: Create or update the company profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompanyInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
  /company/funding:
    get:
      operationId: listFunding
      tags: [company]
      summary: List capital funding entries (own funds + bank loans)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FundingEntry" }
    post:
      operationId: createFunding
      tags: [company]
      summary: Add a funding source
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FundingInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FundingEntry" }
  /company/funding/{id}:
    delete:
      operationId: deleteFunding
      tags: [company]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted

  # ============ FIXED ASSETS ============
...
  /insights/narrate:
    post:
      operationId: narrateInsights
      tags: [insights]
      summary: Explain and prioritise findings the browser already computed
      description: >
        The Rapports page detects findings in the browser — it is client-side by
        convention, and re-deriving its aggregates server-side would create a
        second source of truth. So only the findings cross the wire, never the
        ledger: the API key stays here and the company's rows stay in the browser.

        Gemini receives figures that deterministic code already produced and is
        asked to order and explain them. It is never asked to compute anything —
        the numbers in the narrative are the ones in the request.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InsightNarrateInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsightsNarrative" }
        "503":
          description: >
            Narration is unavailable — no API key, or Gemini declined every
            attempt. The caller already holds the deterministic explanation for
            every finding, so this is expected to degrade to that text rather
            than surface an error.
components:
...
  schemas:
...
    Company:
      type: object
      required: [id, name, nif, ai, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        nif: { type: string, description: "Numéro d'Identification Fiscale" }
        ai: { type: string, description: "Article d'Imposition" }
        address: { type: string, nullable: true }
        legalForm: { type: string, nullable: true, description: "e.g. EURL, SARL, SNC" }
        taxRegime:
          type: string
          nullable: true
          enum: [FORFAITAIRE, REEL]
          description: >-
            Algerian tax system: FORFAITAIRE (النظام الجزافي) files the G12,
            REEL (النظام الحقيقي) files the G50.
        sectorCode:
          type: string
          nullable: true
          description: >-
            Sector of activity, as a code from the @workspace/sectors catalogue.
            Used to filter Journal Officiel decrees down to the ones that
            concern this company. Deliberately not an enum here: the catalogue
            is TypeScript and is the single source of truth for valid codes, so
            an enum duplicated into this file would drift. The server validates
            the code against the catalogue and rejects an unknown one.
        sectorLabel:
          type: string
          nullable: true
          description: >-
            Human-readable sector name, derived server-side from sectorCode.
            Returned only — a client cannot set it, so the stored label can
            never disagree with the code it belongs to.
        createdAt: { type: string, format: date-time }
    CompanyInput:
      type: object
      required: [name, nif, ai]
      properties:
        name: { type: string, minLength: 1 }
        nif: { type: string, minLength: 1 }
        ai: { type: string, minLength: 1 }
        address: { type: string, nullable: true }
        legalForm: { type: string, nullable: true }
        taxRegime:
          type: string
          nullable: true
          enum: [FORFAITAIRE, REEL]
          description: >-
            Algerian tax system: FORFAITAIRE (النظام الجزافي) files the G12,
            REEL (النظام الحقيقي) files the G50.
        sectorCode:
          type: string
          nullable: true
          description: >-
            Sector of activity, as a code from the @workspace/sectors catalogue.
            Validated server-side against that catalogue. `sectorLabel` is
            derived from it and is therefore not accepted on input.
    FundingEntry:
      type: object
      required: [id, source, amount, date]
      properties:
        id: { type: string, format: uuid }
        source: { type: string, enum: [OWN_FUNDS, BANK_LOAN] }
        label: { type: string, nullable: true }
        amount: { type: number }
        date: { type: string, format: date }
        interestRate: { type: number, nullable: true, description: "Annual rate for loans (e.g. 6.5)" }
        durationMonths: { type: integer, nullable: true }
    FundingInput:
      type: object
      required: [source, amount, date]
      properties:
        source: { type: string, enum: [OWN_FUNDS, BANK_LOAN] }
        label: { type: string, nullable: true }
        amount: { type: number, minimum: 0 }
        date: { type: string, format: date }
        interestRate: { type: number, nullable: true }
        durationMonths: { type: integer, nullable: true }
...
    InsightFinding:
      type: object
      description: >
        One observation, as the browser computed it. The field names mirror the
        client's `Finding` exactly, because this is a mirror of that type rather
        than a re-modelling of it.
      required: [ruleId, severity, title, periodRef]
      properties:
        ruleId:
          type: string
          description: >
            The knowledge-base key. Opaque to the server, which only echoes it
            back — the client owns the rules and their wording.
        severity: { type: string, enum: [CRITICAL, WARNING, INFO] }
        title: { type: string }
        evidence:
          type: array
          items: { type: string }
          description: "The figures behind the finding, already formatted for reading"
        metrics:
          type: object
          additionalProperties: { type: number }
          description: "The same figures unformatted — what the narration may reason over"
        periodRef: { type: string }
        subject: { type: string, nullable: true }
    InsightNarrateInput:
      type: object
      required: [findings]
      properties:
        periodFrom: { type: string, format: date, nullable: true }
        periodTo: { type: string, format: date, nullable: true }
        findings:
          type: array
          items: { $ref: "#/components/schemas/InsightFinding" }
    InsightPriority:
      type: object
      required: [ruleId, whyItMatters, action]
      properties:
        ruleId:
          type: string
          description: "Echoes one of the ruleIds in the request — never a new one"
        whyItMatters: { type: string }
        action: { type: string }
    InsightsNarrative:
      type: object
      required: [summary, priorities, model, cached]
      properties:
        summary:
          type: string
          description: "Two or three sentences, in plain French, for a business owner"
        priorities:
          type: array
          items: { $ref: "#/components/schemas/InsightPriority" }
        model: { type: string }
        cached:
          type: boolean
          description: "True when this is a stored narrative for identical findings"